requirements.txt
Augmentor
discord.py~=2.5.0
numpy
aiohttp
loguru~=0.6.0